Add Anticipated Expression for the lazy code motion analysis
**Description**

This commit adds Anticipated Expression analysis, which is
a backward dataflow analysis used as 1st step in the Lazy
Code Motion.

Idea is if an "expression" is antipated in "all" the program
path, then it can be "moved up", and just be computed once
unless re-defined.

Before running anticipated Expressions, we need to
break the critical edges.
Ideally, critical edges those with a basic block
which has multiple successors, and also multiple
predecessors.

**Testing**

TODO: Handle PHI nodes, because its in BB 14, which is the
last BB, since PHI node is defined:
%.1 = PHI ...
Its kills the expression %15 = add nsw i32 %.1, %2

**************************************************
* Instruction-Domain Value Mapping
**************************************************

{}
  %4 = icmp sgt i32 %0, 5

{}
  br i1 %4, label %5, label %7
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}

{[add %1, %2], }
  %6 = add nsw i32 %1, %2

{}
  br label %14
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}

{}
  br label %8
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}

{}
  %.0 = phi i32 [ %1, %7 ], [ %11, %10 ]

{}
  %9 = icmp slt i32 %.0, 5

{}
  br i1 %9, label %10, label %13
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}

{[add %.0, 1], }
  %11 = add nsw i32 %.0, 1

{[add %11, %2], }
  %12 = add nsw i32 %11, %2

{}
  br label %8
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}

{}
  br label %14
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}

{}
  %.1 = phi i32 [ %1, %5 ], [ %.0, %13 ]

{[add %.1, %2], }
  %15 = add nsw i32 %.1, %2

{}
  ret i32 %15
**[Last Inst of BB] -- [Getting Successors: Mask]**
{}