Add support to identify the invariant instruction
**Description**

Add code changes to identify the instruction, that are
Loop Invariant, conditions to check that are:
  1. Definitions are defined outside the loop.
  2. There is only 1 definition, and that is from loop invariant instruction
     within the loop.
  3. Constant/Argument

However, we also have additional conditions

isSafeToSpeculativelyExecute: This checks whether the instruction I can be
executed without side effects.

For Example: **`x = *ptr`** (dereferencing a pointer)
This is unsafe because the pointer might be invalid or point to an unmapped
memory location, leading to a segmentation fault or other memory access violation.

Similarly, mayReadFromMemory This ensures that the instruction does not read from
memory.
Load instructions:
    Example: x = *ptr
    - This reads data from the memory location pointed to by `ptr`. If the value
      at `ptr` changes during each iteration of the loop, moving this instruction
      outside the loop would result in incorrect behavior.

Array access:
   For Example: x = array[i]
    - This reads from an array stored in memory. If different elements of the array
      are accessed in each iteration, moving this outside the loop would result in
      reading the same element repeatedly, which is likely incorrect.