Add AllocationHints implementation for the Virtual Register
**Description**

Register Class
-------------------------------------------------------------------------------
Each Virtual Register (Inst) belongs to a Register Class. Register
class is egisters with similar properties, such as size (e.g., 32-bit
registers) or purpose (e.g., floating-point registers).

For example, on x86, there might be register classes for general-purpose
registers (GPRs), floating-point registers (FPRs), and vector registers (VRs).

During register allocation, when a virtual register needs to be assigned to a
physical register, the allocator can choose any register from the appropriate
register class.

Now based on this Register Class to which virtual register belongs, we get the
current allocation order. This specifies the list of Register order we need
to use for assigning this VirtReg to a PhysReg.

Note: AllocationHints are different from the AllocationOrder.
-------------------------------------------------------------------------------
AllocationHints in LLVM are suggestions for the register allocator about
which physical registers might be preferable for a given virtual register.
Used as input to create an AllocationOrder.

AllocationOrder Encapsulates the preferred allocation order for a virtual register
Takes into account both allocation hints and the target's register allocation order.

AllocationHints are suggestions, while AllocationOrder is a concrete sequence
of physical registers.

However, if we see the input for AllocationHints,it takes in "current allocation
order". This may seem counter intuitive as AllocationOrder takes into account
the Allocation Hint. But this is done for:
- Validation of hints: The allocation order is used to validate that the hints are
within the set of allocatable registers for the given register class. This ensures
that hints don't suggest invalid registers

Hard Hints vs Soft Hints
---------------------------------------------------------------------------------
We get the allocationHints from the "getRegAllocationHints" method which returns a
boolean indicating whether the hints provided are hard hints or not.

If its true, then it is a hard requirement that the hint is followed. Which means
The target-specific code has provided strong preferences for register allocation.

Else, we it is a soft hint, and we usually copy the allocation Order as the hints.