Insert Node to the inference graph, and calculate LI weights
**Descripton**

This commit loops through all the virtual registers, and
identifies all the physical register and virtual registers
it interferes with, and inserts this node in the Infererence
graph.

Inference graph is a map of Live Interval, Set of node it interferes with.

After which we calcuate the weight for the live interval

Cost = [(# defs & uses) * 10^loop-nest-depth] / degree

1. (# defs & uses): This represents the total number of definitions
and uses of the variable. More frequent use of a variable increases
its cost, making it less desirable to spill.

2. 10^loop-nest-depth: This factor accounts for the nesting depth of
loops where the variable is used. Variables used in deeply nested loops
are more expensive to spill because they would require frequent loads
and stores.

3. degree: This represents the number of interferences the variable has
with other variables in the interference graph. A higher degree means the
variable conflicts with more other variables, making it a better candidate
for spilling.